requirements.txt
asgiref==3.8.1
autopep8==2.3.1
Django==4.2.16
django-filter==24.3
djangorestframework==3.15.2
importlib_metadata==8.4.0
Markdown==3.7
mysqlclient==2.2.4
pycodestyle==2.12.1
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
zipp==3.20.1
